Fix output of PV lines with invalid scores #2439
As reported on the forum it is possible, on very rare occasions, that we
are
trying to print a PV line with an invalid previousScore, although this
line
has a valid actual score. This patch fixes output of PV lines with
invalid
scores in a MultiPV search. This is a follow-up patch to 8b15961 and
makes
the fix finally complete.

The reason is the i <= pvIdx condition which probably is a leftover from
the
times there was a special root search function. This check is no longer
needed
today and prevents PV lines past the current one (current pvIdx) to be
flagged
as updated even though they do have a valid score.

8b15961

https://groups.google.com/forum/?fromgroups=#!topic/fishcooking/PrnoDLvMvro

No functional change.
master
@joergoster
@snicolet
joergoster authored